+++ pages/cmp_formulas_data.php
@@ -5,8 +5,8 @@ require(__ROOT__.'/inc/sec.php');
 require_once(__ROOT__.'/inc/config.php');
 require_once(__ROOT__.'/inc/opendb.php');
 
-$id_a = mysqli_real_escape_string($conn, $_GET['id_a']);
-$id_b = mysqli_real_escape_string($conn, $_GET['id_b']);
+$id_a = mysqli_real_escape_string($conn, $_POST['id_a']);
+$id_b = mysqli_real_escape_string($conn, $_POST['id_b']);
 
 
 ?>